Refactor Account and Course context entities into separate factories
